backend/parser.py
import re, docx2txt
from pdfminer.high_level import extract_text
from docx import Document


def extract_text_from_pdf(file_path):
    if file_path.endswith(".docx"):
        doc = Document(file_path)
        text = ""
        for paragraph in doc.paragraphs:
            text += paragraph.text + "\n"
        return text
    elif file_path.endswith(".doc"):
        pass
    return extract_text(file_path)


def extract_contact_number_from_resume(text):
    contact_number = None
    pattern = r"\b(?:\+?\d{1,3}[-.\s]?)?\(?\d{3}\)?[-.\s]?\d{3}[-.\s]?\d{4}\b"
    match = re.search(pattern, text)
    if match:
        contact_number = match.group()
    return contact_number


def extract_email_from_resume(text):
    email = None
    pattern = r"\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}\b"
    match = re.search(pattern, text)
    if match:
        email = match.group()
    return email


def extract_name_from_resume(text):
    name = None
    pattern = r"(\b[A-Z][a-z]+\b)\s(\b[A-Z][a-z]+\b)"
    match = re.search(pattern, text)
    if match:
        name = match.group()
    return name


if __name__ == '__main__':
    text = extract_text_from_pdf(f'Resume.docx')
    print(extract_contact_number_from_resume(text))
    print(extract_email_from_resume(text))
    print(extract_name_from_resume(text))

 
    