backend/requirements.txt
blis==0.7.9
catalogue==1.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
confection==0.1.0
cryptography==41.0.1
cymem==2.0.7
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.0
packaging==23.1
pandas==2.0.3
pathy==0.10.2
pdfminer.six==20221105
pdfplumber==0.9.0
phonenumbers==8.13.15
Pillow==10.0.0
plac==1.1.3
preshed==3.0.8
pycparser==2.21
pydantic==1.10.11
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
resume-parser==0.8.4
six==1.16.0
smart-open==6.3.0
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==1.0.6
stemming==1.0.1
thinc==7.4.6
tika==2.6.0
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
Wand==0.6.11
wasabi==0.10.1